requirements.txt
tensorflow
nltk
spacy
numpy
scikit-learn
Pillow